Don't assume UPDATE rowcount returns matching rows
UPDATE queries can return either matching or changed rows. The
connection code in the db submodule does not explicitly set the client
option for this and at some point the default must have changed leading
to incorrect errors.

By performing a SELECT first and checking the rowcount of that query, we
can be sure we're looking at the right result.